token checked before expired

+++ MyBlogFront/src/app/services/interceptors/token.interceptor.ts
@@ -13,7 +13,7 @@ export class TokenInterceptor implements HttpInterceptor {
     intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
         const token = localStorage.getItem('token');
 
-        if (this.jwtService.isTokenExpired(token)) {
+        if (token !== null && this.jwtService.isTokenExpired(token)) {
             localStorage.clear();
             this.route.navigate(['/']);
         }
